--- src/classes/Listings/export.ts
@@ -1,7 +1,0 @@
-import replaceSpells from './details/spells';
-import replaceKillstreaker from './details/killstreakers';
-import replaceSheens from './details/sheens';
-import replacePainted from './details/painted';
-import replaceStrangeParts from './details/strangeParts';
-
-export { replaceSpells, replaceKillstreaker, replaceSheens, replacePainted, replaceStrangeParts };
